Photo configuration

+++ app/Http/Controllers/Admin/ConfigurationPhotoController.php
@@ -12,25 +12,28 @@ class ConfigurationPhotoController extends Controller
 {
     public function store(Request $request, $id)
     {
-        $imageName = Str::random(10) . '.' . 'jpg';  
+        $imageName = Str::random(10) . '.' . 'jpg';
         $photo = new ConfigurationPhoto();
         $photo->configuration_id = $id;
         $photo->filename = $imageName;
         $photo->save();
-        
-        $request->file('file')->storeAs('images/configuration/covers/'.$id.'/', $imageName, 'public');
 
-        return response()->json(["msg"=>"Photo added succesfully"], 200);
+        $request->file('file')->storeAs('images/configuration/covers/' . $id . '/', $imageName, 'public');
+
+        return response()->json(["msg" => "Photo added succesfully"], 200);
     }
 
     public function destroy(Request $request, $id, $photo_id)
     {
         $photo = ConfigurationPhoto::find($photo_id);
-        if (\File::exists(public_path('images/configuration/covers/1/' . $photo->filename))){
-            \File::delete(public_path('images/configuration/covers/1/' . $photo->filename));
+        if (!$photo) {
+            return response()->json(["msg" => "No se encontro la foto"], 400);
+        } else {
+            if (\File::exists(public_path('images/configuration/covers/1/' . $photo->filename))) {
+                \File::delete(public_path('images/configuration/covers/1/' . $photo->filename));
+            }
+            $photo->delete();
+            return response()->json(["data" => $photo], 200);
         }
-        $photo->delete();
-
-        return redirect()->back()->with('success', 'Foto eliminada correctamente.');
     }
 }
